+++ apps/alpha-test-app/src/components/layout-demo/index.ts
@@ -8,7 +8,6 @@ export { default as MobileBottomFlyInDemo } from "./MobileBottomFlyInDemo.vue";
 export { default as MobileNavFlyoutDemo } from "./MobileNavFlyoutDemo.vue";
 export { default as NavBarDemo } from "./NavBarDemo.vue";
 export { default as PageDemo } from "./PageDemo.vue";
-export { default as PopoverDemo } from "./PopoverDemo.vue";
 export { default as SidebarDemo } from "./SidebarDemo.vue";
 export { default as StickyDemo } from "./StickyDemo.vue";
 export { default as TempOverlayDemo } from "./TempOverlayDemo.vue";