+++ packages/sit-onyx/src/components/OnyxNavBar/types.ts
@@ -1,3 +1,4 @@
+import type { OnyxBreakpoint } from "../../types";
 import type { OnyxNavAppAreaProps } from "../OnyxNavAppArea/types";
 
 export type OnyxNavBarProps = OnyxNavAppAreaProps & {
@@ -10,4 +11,12 @@ export type OnyxNavBarProps = OnyxNavAppAreaProps & {
    * @default Translated "Go to home" depending on the current locale.
    */
   appAreaLabel?: string;
+  /**
+   * Breakpoint name when the nav bar should switch into mobile mode.
+   * Will switch if smaller or equal to the given breakpoint.
+   * Can be either a pre-defined onyx breakpoint or custom width in pixels.
+   *
+   * @see [onyx docs](https://onyx.schwarz/development/breakpoints.html) for more information.
+   */
+  mobileBreakpoint?: OnyxBreakpoint | number;
 };