+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @sit-onyx/nuxt
 
+## 1.0.0-alpha.17
+
+### Patch Changes
+
+- Updated dependencies [442e4c5]
+  - sit-onyx@1.0.0-alpha.137
+
+## 1.0.0-alpha.16
+
+### Patch Changes
+
+- Updated dependencies [f1aad40]
+  - sit-onyx@1.0.0-alpha.136
+
+## 1.0.0-alpha.15
+
+### Patch Changes
+
+- Updated dependencies [0511127]
+  - sit-onyx@1.0.0-alpha.135
+
+## 1.0.0-alpha.14
+
+### Patch Changes
+
+- Updated dependencies [02d9f0f]
+  - sit-onyx@1.0.0-alpha.134
+
+## 1.0.0-alpha.13
+
+### Patch Changes
+
+- sit-onyx@1.0.0-alpha.133
+
+## 1.0.0-alpha.12
+
+### Patch Changes
+
+- Updated dependencies [6059d12]
+  - sit-onyx@1.0.0-alpha.132
+
+## 1.0.0-alpha.11
+
+### Patch Changes
+
+- Updated dependencies [1baef56]
+  - sit-onyx@1.0.0-alpha.131
+
+## 1.0.0-alpha.10
+
+### Patch Changes
+
+- Updated dependencies [37ee3fd]
+  - sit-onyx@1.0.0-alpha.130
+
 ## 1.0.0-alpha.9
 
 ### Patch Changes