+++ packages/storybook-utils/CHANGELOG.md
@@ -1,5 +1,68 @@
 # @sit-onyx/storybook-utils
 
+## 1.0.0-alpha.142
+
+### Patch Changes
+
+- Updated dependencies [442e4c5]
+  - sit-onyx@1.0.0-alpha.137
+
+## 1.0.0-alpha.141
+
+### Patch Changes
+
+- Updated dependencies [f1aad40]
+  - sit-onyx@1.0.0-alpha.136
+
+## 1.0.0-alpha.140
+
+### Patch Changes
+
+- Updated dependencies [0511127]
+  - sit-onyx@1.0.0-alpha.135
+
+## 1.0.0-alpha.139
+
+### Patch Changes
+
+- Updated dependencies [02d9f0f]
+  - sit-onyx@1.0.0-alpha.134
+
+## 1.0.0-alpha.138
+
+### Patch Changes
+
+- Updated dependencies [fad8140]
+  - @sit-onyx/icons@0.1.0-alpha.2
+  - sit-onyx@1.0.0-alpha.133
+
+## 1.0.0-alpha.137
+
+### Patch Changes
+
+- Updated dependencies [6059d12]
+  - sit-onyx@1.0.0-alpha.132
+
+## 1.0.0-alpha.136
+
+### Patch Changes
+
+- Updated dependencies [1baef56]
+  - sit-onyx@1.0.0-alpha.131
+
+## 1.0.0-alpha.135
+
+### Patch Changes
+
+- Updated dependencies [37ee3fd]
+  - sit-onyx@1.0.0-alpha.130
+
+## 1.0.0-alpha.134
+
+### Minor Changes
+
+- 48dedc3: feat: add imports for used onyx components to generated source code
+
 ## 1.0.0-alpha.133
 
 ### Patch Changes